feature(signup) authenticate password [fixes #167121532]
