+++ .travis.yml
@@ -1,10 +1,14 @@
 language: node_js
 node_js:
 - "stable"
+services:
+- postgresql
 install: 
 - npm install
 cache:
   directories:
     - node_modules
-script:
-  - npm run lint
+script: 
+- npm test
+after_success: 
+- npm run test-coverage
