+++ server/config/config.js
@@ -1,5 +1,3 @@
-//separate different environments
-
 import dotenv from 'dotenv';
 
 dotenv.config();
