+++ server/helpers/token.js
