+++ server/routes/index.js
@@ -1,7 +1,6 @@
 import express from 'express';
-
 import {validateSingupCredentials} from '../middlewares/validate';
-import User from '../controllers/users';
+import User from '../controllers/User';
 
 const router = express.Router();
